Add Test Coverage To CI
This patch adds code overage to the CI pipeline, giving the developers
of the project the ability to know how much of their code is tested.
For our project, we should be looking to meet a minimum level of 90%
coverage for a given file.

This patch also cleans up the requirements files to only have the
primary dependencies listed.