Adds initial app structure
To keep things simple, we're starting with a single shared Django
app and will expand as needed.  This initial app structure includes
the following features:

- app directory structure
- app added to installed apps in settings
- placeholder templates
- placeholder static files (css)
- development static file hosting
- tests directory
- core.urls added to gradify.urls

To ensure this is working on your local machine, run ./manage.py runserver
and visit the root of the web application in the browser to see a
page now serving.