Tunes CI pipeline and adds a script for local testing
Adjusts the CI configuration to ignore directories which we are
not directly working with. This should speed up the CI a bit and
prevent linter complaints about files created by Django. Note that
according to Github issues fiiled, bandit's exclude functionality
appears to be broken in version 1.6.0. Changed it to only scan
./core and ./gradify.

Additionally, includes a script for running the CI pipeline jobs
locally to test things before pushing changes upstream.