Adds script for loading fixtures
Applies any new migrations and loads all fixtures found in `gradify/core/fixtures`
Turns out that if you pass all the fixtures at once as arguments to loaddata,
Django will do the work of figuring out what order they go in. This allows
scripting it to get the contents of the fixtures dir, strip the file extensions,
and pass the list as arguments to loaddata. Anytime a new fixture as added,
just rerun this script after pulling in the latest migrations and you're g2g.

Closes #36