Refactors gradebook logic
Builds on the work done to display averages in the gradebook.
Rather than add additonal loops and reduce efficiency, the
pre-existing for loop on the submissions is used to prepare
the values necessary to calculate the averages.

Only submissions that have an assigned_grade are factored into
each student's average, regardless of due date.