Updates bandit CI configuration
Sets bandit to ignore B105 - Hardcoded password string.
While the secret key should not be stored in the settings file
in a production environment, taking it out during development
will create some friction. Issue #49 has been created to ensure
that this is just a temporary measure to clean up the CI output.
To serve as an additonal reminder, the `run_pipeline.sh` script
does not ignore the rule when run locally.

Closes #43