Adds ingest of courses from Google Classroom
Begins implementing Google Classroom ingest. Several changes
were required:
* An `owner` FK has been added to the Course model to denote the
user that imported the Google Classroom course
* User "teacher1" password has been changed to "password". This
was necessary because the course list now filters for courses
where the owner is the currently logged in user, so the tests
had to be modified to log in with credentials.
* To simplify Google Classroom ingest, new fields for the GC fields
`creationTime` and `updateTime` were created in the `Course` model
* A classmethod, `create_gc_course` was created to simplify ingest.
This method deletes unneeded dict items and passes the dict as kwargs
* `__str__` methods were added to those models that did not have one
so that interpreting data on the admin site is easier

Also updates the Gradebook to display all students in the course,
rather than just those that have submitted assignments. This would
have been a separate commit, but with all the changes, the migrations
needed to be blown away and rebuilt so it was unclear if things would
work.

Supports #9