Refactors Gradebook
The previous implementation of gradebook used more queries and
calculation than necessary.  this refactor also adds the ability
to add a grade for a given student_submission without using drop
downs on the form.

Note: this patch was meant to demonstrate a different approach to
what had already been done.  I didn't waste time on fixing the unit
tests, because I wasn't sure if this code would be kept, used as
inspiriation for another patch or just discarded.